requirements.txt
python-can>=4.0.0
canopen>=2.0.0
pytest>=7.0.0
